+++ ssc/cmod_iscc_design_point.cpp
@@ -103,9 +103,9 @@ public:
 				mat.assign(htf_mat, nrows, ncols);
 
 				util::matrix_t<double> mat_double(nrows, ncols);
-				for( int i = 0; i < nrows; i++ )
+				for( size_t i = 0; i < nrows; i++ )
 				{
-					for( int j = 0; j < ncols; j++ )
+					for( size_t j = 0; j < ncols; j++ )
 					{
 						mat_double(i, j) = (double)mat(i, j);
 					}
