+++ tcs/csp_solver_pc_gen.cpp
@@ -280,9 +280,9 @@ void C_pc_gen::call(const C_csp_weatherreader::S_outputs &weather,
 	// Calculate the load-based and temperature-based efficiency correction factors
 	double f_effpc_qtpb = 0.0;
 	double f_effpc_tamb = 0.0;
-	for( int i = 0; i < ms_params.mv_etaQ_coefs.size(); i++ )
+	for( size_t i = 0; i < ms_params.mv_etaQ_coefs.size(); i++ )
 		f_effpc_qtpb += ms_params.mv_etaQ_coefs[i] * pow(qnorm, i);
-	for( int i = 0; i < ms_params.mv_etaT_coefs.size(); i++ )
+	for( size_t i = 0; i < ms_params.mv_etaT_coefs.size(); i++ )
 		f_effpc_tamb += ms_params.mv_etaT_coefs[i] * pow(tnorm, i);
 	
 	double eta_cycle = ms_params.m_eta_des * (f_effpc_qtpb + f_effpc_tamb);  //[-] Adjusted power conversion efficiency
